Split off the application instance search view
