+++ lms/views/admin/application_instance/view.py
@@ -12,7 +12,6 @@ from lms.services import ApplicationInstanceNotFound, LTIRegistrationService
 from lms.services.aes import AESService
 from lms.validation._base import PyramidRequestSchema, ValidationError
 from lms.views.admin import flash_validation
-from lms.views.admin._schemas import EmptyStringInt
 
 # Helper to declare settings as secret
 AES_SECRET = object()
@@ -82,19 +81,6 @@ class UpgradeApplicationInstanceSchema(PyramidRequestSchema):
 
     consumer_key = fields.Str(required=True, validate=validate.Length(min=1))
     deployment_id = fields.Str(required=True, validate=validate.Length(min=1))
-
-
-class SearchApplicationInstanceSchema(PyramidRequestSchema):
-    location = "form"
-
-    # Max value for postgres `integer` type
-    id = EmptyStringInt(required=False, validate=validate.Range(max=2147483647))
-    name = fields.Str(required=False)
-    consumer_key = fields.Str(required=False)
-    issuer = fields.Str(required=False)
-    client_id = fields.Str(required=False)
-    deployment_id = fields.Str(required=False)
-    tool_consumer_instance_guid = fields.Str(required=False)
 
 
 @view_defaults(request_method="GET", permission=Permissions.ADMIN)
@@ -257,53 +243,6 @@ class AdminApplicationInstanceViews:
             self.request.session.flash("Downgraded LTI 1.1 successful", "messages")
 
         return self._redirect("admin.instance", id_=ai.id)
-
-    @view_config(
-        route_name="admin.instance.search",
-        renderer="lms:templates/admin/application_instance/search.html.jinja2",
-    )
-    def search_start(self):
-        return {"settings": APPLICATION_INSTANCE_SETTINGS_COLUMNS}
-
-    @view_config(
-        route_name="admin.instance.search",
-        request_method="POST",
-        require_csrf=True,
-        renderer="lms:templates/admin/application_instance/search.html.jinja2",
-    )
-    def search_callback(self):
-        if flash_validation(self.request, SearchApplicationInstanceSchema):
-            return {}
-
-        settings = None
-        if settings_key := self.request.params.get("settings_key"):
-            if settings_value := self.request.params.get("settings_value"):
-                settings_value = APPLICATION_INSTANCE_SETTINGS.get(
-                    tuple(settings_key.split("."))
-                )(settings_value)
-            else:
-                settings_value = ...
-
-            settings = {settings_key: settings_value}
-
-        instances = self.application_instance_service.search(
-            id_=self.request.params.get("id"),
-            name=self.request.params.get("name"),
-            consumer_key=self.request.params.get("consumer_key"),
-            issuer=self.request.params.get("issuer"),
-            client_id=self.request.params.get("client_id"),
-            deployment_id=self.request.params.get("deployment_id"),
-            tool_consumer_instance_guid=self.request.params.get(
-                "tool_consumer_instance_guid"
-            ),
-            email=self.request.params.get("email"),
-            settings=settings,
-        )
-
-        return {
-            "instances": instances,
-            "settings": APPLICATION_INSTANCE_SETTINGS_COLUMNS,
-        }
 
     @view_config(
         route_name="admin.instance",
